improved code to increase readability
